All CRUD Operations implemented using Hibernate

+++ src/main/java/com/example/JspStudentCrud/DB/StudentDaoHbnt.java
@@ -55,7 +55,7 @@ public class StudentDaoHbnt {
      *
      * @param id
      */
-    public void deleteStudent(int id) {
+    public void deleteStudent(Long id) {
         Transaction transaction = null;
         try (Session session = HibernateUtil.getSessionFactory().openSession()) {
             // start a transaction
@@ -81,7 +81,7 @@ public class StudentDaoHbnt {
      * @param id
      * @return
      */
-    public Student getStudent(int id) {
+    public Student getStudent(Long id) {
         Transaction transaction = null;
         Student student = null;
         try (Session session = HibernateUtil.getSessionFactory().openSession()) {
